signa-backend/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-environ==0.11.2
django-cors-headers==4.3.1
django-filter==24.2
django-redis==5.4.0
celery==5.4.0
redis==5.0.4
ephem==4.1.5
psycopg2-binary==2.9.9
gunicorn==22.0.0
stripe==9.5.0
Pillow==10.3.0
whitenoise==6.6.0
